Lift filters parsing and list parsing out of QueryParser
