+++ lib/plausible/stats/filters/query_parser.ex
@@ -2,6 +2,8 @@ defmodule Plausible.Stats.Filters.QueryParser do
   @moduledoc false
 
   alias Plausible.Stats.{TableDecider, Filters, Metrics, DateTimeRange, JSONSchema}
+  alias Plausible.Stats.Filters.FiltersParser
+  alias Plausible.Helpers.ListTraverse
 
   @default_include %{
     imports: false,
@@ -28,7 +30,7 @@ defmodule Plausible.Stats.Filters.QueryParser do
            parse_time_range(site, Map.get(params, "date_range"), date, now),
          utc_time_range = raw_time_range |> DateTimeRange.to_timezone("Etc/UTC"),
          {:ok, metrics} <- parse_metrics(Map.get(params, "metrics", [])),
-         {:ok, filters} <- parse_filters(Map.get(params, "filters", [])),
+         {:ok, filters} <- FiltersParser.parse_filters(Map.get(params, "filters", [])),
          {:ok, dimensions} <- parse_dimensions(Map.get(params, "dimensions", [])),
          {:ok, order_by} <- parse_order_by(Map.get(params, "order_by")),
          {:ok, include} <- parse_include(Map.get(params, "include", %{})),
@@ -57,7 +59,7 @@ defmodule Plausible.Stats.Filters.QueryParser do
   end
 
   defp parse_metrics(metrics) when is_list(metrics) do
-    parse_list(metrics, &parse_metric/1)
+    ListTraverse.parse_list(metrics, &parse_metric/1)
   end
 
   defp parse_metric(metric_str) do
@@ -66,89 +68,6 @@ defmodule Plausible.Stats.Filters.QueryParser do
       _ -> {:error, "Unknown metric '#{i(metric_str)}'."}
     end
   end
-
-  def parse_filters(filters) when is_list(filters) do
-    parse_list(filters, &parse_filter/1)
-  end
-
-  def parse_filters(_invalid_metrics), do: {:error, "Invalid filters passed."}
-
-  defp parse_filter(filter) do
-    with {:ok, operator} <- parse_operator(filter),
-         {:ok, second} <- parse_filter_second(operator, filter),
-         {:ok, rest} <- parse_filter_rest(operator, filter) do
-      {:ok, [operator, second | rest]}
-    end
-  end
-
-  defp parse_operator(["is" | _rest]), do: {:ok, :is}
-  defp parse_operator(["is_not" | _rest]), do: {:ok, :is_not}
-  defp parse_operator(["matches" | _rest]), do: {:ok, :matches}
-  defp parse_operator(["matches_not" | _rest]), do: {:ok, :matches_not}
-  defp parse_operator(["matches_wildcard" | _rest]), do: {:ok, :matches_wildcard}
-  defp parse_operator(["matches_wildcard_not" | _rest]), do: {:ok, :matches_wildcard_not}
-  defp parse_operator(["contains" | _rest]), do: {:ok, :contains}
-  defp parse_operator(["contains_not" | _rest]), do: {:ok, :contains_not}
-  defp parse_operator(["not" | _rest]), do: {:ok, :not}
-  defp parse_operator(["and" | _rest]), do: {:ok, :and}
-  defp parse_operator(["or" | _rest]), do: {:ok, :or}
-  defp parse_operator(filter), do: {:error, "Unknown operator for filter '#{i(filter)}'."}
-
-  def parse_filter_second(:not, [_, filter | _rest]), do: parse_filter(filter)
-
-  def parse_filter_second(operator, [_, filters | _rest]) when operator in [:and, :or],
-    do: parse_filters(filters)
-
-  def parse_filter_second(_operator, filter), do: parse_filter_key(filter)
-
-  defp parse_filter_key([_operator, filter_key | _rest] = filter) do
-    parse_filter_key_string(filter_key, "Invalid filter '#{i(filter)}")
-  end
-
-  defp parse_filter_key(filter), do: {:error, "Invalid filter '#{i(filter)}'."}
-
-  defp parse_filter_rest(operator, filter)
-       when operator in [
-              :is,
-              :is_not,
-              :matches,
-              :matches_not,
-              :matches_wildcard,
-              :matches_wildcard_not,
-              :contains,
-              :contains_not
-            ],
-       do: parse_clauses_list(filter)
-
-  defp parse_filter_rest(operator, _filter)
-       when operator in [:not, :and, :or],
-       do: {:ok, []}
-
-  defp parse_clauses_list([operation, filter_key, list] = filter) when is_list(list) do
-    all_strings? = Enum.all?(list, &is_binary/1)
-    all_integers? = Enum.all?(list, &is_integer/1)
-
-    case {filter_key, all_strings?} do
-      {"visit:city", false} when all_integers? ->
-        {:ok, [list]}
-
-      {"visit:country", true} when operation in ["is", "is_not"] ->
-        if Enum.all?(list, &(String.length(&1) == 2)) do
-          {:ok, [list]}
-        else
-          {:error,
-           "Invalid visit:country filter, visit:country needs to be a valid 2-letter country code."}
-        end
-
-      {_, true} ->
-        {:ok, [list]}
-
-      _ ->
-        {:error, "Invalid filter '#{i(filter)}'."}
-    end
-  end
-
-  defp parse_clauses_list(filter), do: {:error, "Invalid filter '#{i(filter)}'"}
 
   defp parse_date(_site, date_string, _date) when is_binary(date_string) do
     case Date.from_iso8601(date_string) do
@@ -258,14 +177,14 @@ defmodule Plausible.Stats.Filters.QueryParser do
   defp today(site), do: DateTime.now!(site.timezone) |> DateTime.to_date()
 
   defp parse_dimensions(dimensions) when is_list(dimensions) do
-    parse_list(
+    ListTraverse.parse_list(
       dimensions,
       &parse_dimension_entry(&1, "Invalid dimensions '#{i(dimensions)}'")
     )
   end
 
   defp parse_order_by(order_by) when is_list(order_by) do
-    parse_list(order_by, &parse_order_by_entry/1)
+    ListTraverse.parse_list(order_by, &parse_order_by_entry/1)
   end
 
   defp parse_order_by(nil), do: {:ok, nil}
@@ -281,7 +200,7 @@ defmodule Plausible.Stats.Filters.QueryParser do
   defp parse_dimension_entry(key, error_message) do
     case {
       parse_time(key),
-      parse_filter_key_string(key)
+      FiltersParser.parse_filter_key_string(key)
     } do
       {{:ok, time}, _} -> {:ok, time}
       {_, {:ok, filter_key}} -> {:ok, filter_key}
@@ -293,7 +212,7 @@ defmodule Plausible.Stats.Filters.QueryParser do
     case {
       parse_time(value),
       parse_metric(value),
-      parse_filter_key_string(value)
+      FiltersParser.parse_filter_key_string(value)
     } do
       {{:ok, time}, _, _} -> {:ok, time}
       {_, {:ok, metric}, _} -> {:ok, metric}
@@ -323,34 +242,6 @@ defmodule Plausible.Stats.Filters.QueryParser do
 
   defp atomize_keys(map),
     do: Map.new(map, fn {key, value} -> {String.to_existing_atom(key), value} end)
-
-  defp parse_filter_key_string(filter_key, error_message \\ "") do
-    case filter_key do
-      "event:props:" <> property_name ->
-        if String.length(property_name) > 0 do
-          {:ok, filter_key}
-        else
-          {:error, error_message}
-        end
-
-      "event:" <> key ->
-        if key in Filters.event_props() do
-          {:ok, filter_key}
-        else
-          {:error, error_message}
-        end
-
-      "visit:" <> key ->
-        if key in Filters.visit_props() do
-          {:ok, filter_key}
-        else
-          {:error, error_message}
-        end
-
-      _ ->
-        {:error, error_message}
-    end
-  end
 
   defp validate_order_by(query) do
     if query.order_by do
@@ -423,7 +314,10 @@ defmodule Plausible.Stats.Filters.QueryParser do
       end)
 
     if length(goal_filter_clauses) > 0 do
-      validate_list(goal_filter_clauses, &validate_goal_filter(&1, query.preloaded_goals))
+      ListTraverse.validate_list(
+        goal_filter_clauses,
+        &validate_goal_filter(&1, query.preloaded_goals)
+      )
     else
       :ok
     end
@@ -467,7 +361,7 @@ defmodule Plausible.Stats.Filters.QueryParser do
   end
 
   defp validate_metrics(query) do
-    with :ok <- validate_list(query.metrics, &validate_metric(&1, query)) do
+    with :ok <- ListTraverse.validate_list(query.metrics, &validate_metric(&1, query)) do
       validate_no_metrics_filters_conflict(query)
     end
   end
@@ -528,22 +422,4 @@ defmodule Plausible.Stats.Filters.QueryParser do
   end
 
   defp i(value), do: inspect(value, charlists: :as_lists)
-
-  defp parse_list(list, parser_function) do
-    Enum.reduce_while(list, {:ok, []}, fn value, {:ok, results} ->
-      case parser_function.(value) do
-        {:ok, result} -> {:cont, {:ok, results ++ [result]}}
-        {:error, _} = error -> {:halt, error}
-      end
-    end)
-  end
-
-  defp validate_list(list, parser_function) do
-    Enum.reduce_while(list, :ok, fn value, :ok ->
-      case parser_function.(value) do
-        :ok -> {:cont, :ok}
-        {:error, _} = error -> {:halt, error}
-      end
-    end)
-  end
 end
